backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pydantic-ai>=0.0.14",
    "pydantic-ai-litellm>=0.0.1",
    "logfire[fastapi]>=0.30.0",
    "rich>=13.0.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.128.0",
    "litellm>=1.77.2",
    "mem0ai>=0.1.116",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.40.0",
    "isort>=7.0.0",
    "black>=26.1.0",
    "pre-commit>=4.5.1",
    "tavily-python>=0.5.0",
    "motor",
    "passlib[bcrypt]",
    "python-multipart",
    "pydantic[email]",
    "slowapi",
    "bcrypt==3.2.2",
    "itsdangerous>=2.2.0",
    "langchain-core",
    "langchain-community",
    "kreuzberg>=4.0.0",
    "pytest>=9.0.2",
    "python-jose[cryptography]>=3.5.0",
    "pyjwt>=2.11.0",
    "websocket>=0.2.1",
    "pandas>=3.0.1",
    "openpyxl>=3.1.5",
    "ultralytics>=8.0.0",
    "opencv-python-headless>=4.5.0",
    "torch>=1.7.0",
    "torchvision>=0.8.0",
    "pyyaml>=5.0.0",
    "redis>=5.0.0",
    "boto3>=1.42.77",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
